% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_adsb}
\alias{calc_adsb}
\title{Calculate ADSB (avg diff in SSB) for use in determining model
names. If ADSB>0.1 then it's considered a major version.}
\usage{
calc_adsb(base, new)
}
\arguments{
\item{base}{The base model std df or rep list}

\item{new}{The propsed model std df or rep list. Can have more
years than base model}
}
\value{
ADSB value and message
}
\description{
Calculate ADSB (avg diff in SSB) for use in determining model
names. If ADSB>0.1 then it's considered a major version.
}
