% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admb_fns.R
\name{clean_pk_dir}
\alias{clean_pk_dir}
\title{Clean temporary files from a directory containing a model run}
\usage{
clean_pk_dir(path = getwd(), full = FALSE)
}
\arguments{
\item{path}{Path to directory to clean}

\item{full}{Whether to delete everything except .dat and .tpl files}
}
\value{
Nothing
}
\description{
Clean temporary files from a directory containing a model run
}
