% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb_fns.R
\name{fit_drop_surveys}
\alias{fit_drop_surveys}
\title{Turn off data components one at a time to see the effect}
\usage{
fit_drop_surveys(
  fits,
  plot = TRUE,
  Nmult = 0,
  CVmult = 10,
  return.fits = FALSE,
  ...
)
}
\arguments{
\item{fits}{A fitted object. Currently does not support a list of fits.}

\item{plot}{Whether to plot it}

\item{Nmult}{How much to multiply the ISS by (<1 to downweight it)}

\item{CVmult}{How much to multiple the index CVs by (>1 to downweight it)}

\item{return.fits}{Whether to return the fits, if FALSE it returns the ggplot}

\item{...}{Further arguments passed to \code{plot_ssb}.}
}
\value{
Either a ggplot object of a list of fits.
}
\description{
Turn off data components one at a time to see the effect
}
\details{
Needs to be updated when the model accepts CVs and multN of zero for all surveys.
}
