% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb_fns.R
\name{fit_pk}
\alias{fit_pk}
\title{Fit a GOA pollock model (BETA)}
\usage{
fit_pk(
  input,
  getsd = TRUE,
  newtonsteps = 1,
  control = NULL,
  do.fit = TRUE,
  use_bounds = FALSE
)
}
\arguments{
\item{input}{Input list as returned by
\code{prepare_pk_input}.}

\item{getsd}{Calculate and return sdreport?}

\item{newtonsteps, control}{Optimizer settings}

\item{do.fit}{Optimize or return obj? Used for testing.}

\item{use_bounds}{Whether to use bounds, slows optimization
down but may be necessary. See \link{\code{get_bounds}}
which is called internally.}
}
\value{
A list object of class 'pkfit' which contains a
"version" model name, rep, opt as returned by
\code{TMBHelper::fit_tmb} but without the SD slot, std
(formatted data frame) and sdrep if \code{getsd=TRUE}, and
the obj.
}
\description{
Fit a GOA pollock model (BETA)
}
\details{
This function is beta still.
}
