% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retro_fns.R
\name{fit_retros}
\alias{fit_retros}
\title{Fit a sequence of models to peeled data set for retrospective
calculations}
\usage{
fit_retros(fit, peels = 0:7, getsd = TRUE, parallel = FALSE, ...)
}
\arguments{
\item{peels}{Vector of peels to fit, with 0 being the original
data set}

\item{getsd}{Whether to run sdreport}

\item{parallel}{Whether to run peels in parallel or not (default).}

\item{...}{Additional arguments to pass to \code{\link{fit_pk}}}

\item{obj}{A fitted model from \code{fit_pk}.}
}
\value{
A list of model fits of class 'pkfit'
}
\description{
Fit a sequence of models to peeled data set for retrospective
calculations
}
\details{
This function fits a series of models to peels based
off an original fit. The data, parameter, and map lists are
modified based on the peel in a way that the original MLEs
are preserved so that it runs faster. Pass an unfitted obj
if this is undesired behavior.
}
