% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb_fns.R
\name{fit_self_test}
\alias{fit_self_test}
\title{Perform a self-test estimation for a fitted model.}
\usage{
fit_self_test(fit, reps, parallel = TRUE)
}
\arguments{
\item{fit}{A fitted object}

\item{reps}{A vector of integers of reps to run}

\item{parallel}{Whether to run in parallel (default)}
}
\value{
A data.frame containing parameter true, estimates,  absolute and relative errors, and maxgrad.
}
\description{
Perform a self-test estimation for a fitted model.
}
