% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_fns.R
\name{format_exec_table}
\alias{format_exec_table}
\title{Format the executive table}
\usage{
format_exec_table(tab)
}
\arguments{
\item{tab}{Table as read in by \link{get_exec_table}}
}
\description{
Format the executive table
}
