% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb_fns.R
\name{get_bounds}
\alias{get_bounds}
\title{Calculate bounds for parameters}
\usage{
get_bounds(obj)
}
\arguments{
\item{obj}{Object}
}
\value{
list of lower (lwr) and upper (upr) vectors matching
the object
}
\description{
Calculate bounds for parameters
}
