% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_fns.R
\name{get_fabc}
\alias{get_fabc}
\title{Calculate Fabc for a fitted object.}
\usage{
get_fabc(fit, ratio = 0.4, F = NULL)
}
\arguments{
\item{fit}{Fitted TMB model}

\item{ratio}{The desired target biomass ratio}

\item{F}{An option fishing mortality rate which can be
specified to get SPR(F), used for status figure}
}
\value{
A list containing SSB100, fabc, abc, btarget
(optimzed) and SPR at F=0, the optimized target, and what is specified by F.
}
\description{
Calculate Fabc for a fitted object.
}
