% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_fns.R
\name{get_nll_components}
\alias{get_nll_components}
\title{Extract individual negative log likelihood components (NLL) in a convenient data.frame}
\usage{
get_nll_components(fits)
}
\arguments{
\item{fits}{A single model or list of models as returned by
\code{\link{fit_pk}}.
@return A data frame of fitted NLL components and their names
@export}
}
\description{
Extract individual negative log likelihood components (NLL) in a convenient data.frame
}
