% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb_fns.R
\name{get_rep}
\alias{get_rep}
\title{Extract report list from fitted models}
\usage{
get_rep(fits, slot = NULL)
}
\arguments{
\item{fits}{A single model or list of models as returned by
\code{\link{fit_pk}}}

\item{slot}{Which slot to extract and format into data.frame
for easy plotting. If NULL (default) the whole report file
is returned.}
}
\value{
Either a list of reports or a formatted data.frame
containing the specified slot variable.
}
\description{
Extract report list from fitted models
}
