% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_fns.R
\name{get_sdnr}
\alias{get_sdnr}
\title{Calculate SDNR for indices}
\usage{
get_sdnr(fit)
}
\arguments{
\item{fit}{A fit or list of fits as returned by
\code{fit_pk}. Not currently implemented for the ADMB model.}
}
\value{
A data.frame of SDNR for all surveys with a column for
version
}
\description{
Calculate SDNR for indices
}
\details{
SDNR=standard deviation of normalized residuals sd((o-e)/std)
}
