% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb_fns.R
\name{get_std}
\alias{get_std}
\title{Extract sdreport data.frames from fitted models}
\usage{
get_std(fits)
}
\arguments{
\item{fits}{A single model or list of models as returned by
\code{\link{fit_pk}}}
}
\value{
A cleanly formatted data frame of sdreport objects,
with columns 'name', 'est', 'se', 'year', 'lwr', 'upr', and
'version'
}
\description{
Extract sdreport data.frames from fitted models
}
