% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.pkfits}
\alias{is.pkfits}
\title{Check if an object is a list of pkfit objects}
\usage{
is.pkfits(x)
}
\arguments{
\item{x}{List of fits returned from \code{\link{fit_tmb}}}
}
\description{
Check if an object is a list of pkfit objects
}
