% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mapoff}
\alias{mapoff}
\title{Map off a parameter in a given input list}
\usage{
mapoff(x, slots)
}
\arguments{
\item{x}{Input list}

\item{slots}{A character vector to map off}
}
\value{
The modified input list
}
\description{
Map off a parameter in a given input list
}
