% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pk_rep.R
\name{mymelt}
\alias{mymelt}
\title{Extract and melt a named object from a replist (beta)}
\usage{
mymelt(replist, slot)
}
\arguments{
\item{replist}{Output from \code{read_pk_rep}. Can be a single
report object or a list of them.}

\item{slot}{The slot name to extract}
}
\value{
A data frame with named arguments sometimes
}
\description{
Extract and melt a named object from a replist (beta)
}
