% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pkfit}
\alias{pkfit}
\title{Constructor for the "pkfit" (A-D fit) class}
\usage{
pkfit(x)
}
\arguments{
\item{x}{Fitted object from \code{\link{fit_tmb}}}
}
\value{
An object of class "pkfit"
}
\description{
Constructor for the "pkfit" (A-D fit) class
}
