% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fns.R
\name{plot_cohort_biomass}
\alias{plot_cohort_biomass}
\title{Plot cohort contribution to SSB as a stacked ribbon plot}
\usage{
plot_cohort_biomass(
  fit,
  type = c("ssb", "total", "summary"),
  plot = TRUE,
  print = FALSE
)
}
\arguments{
\item{fit}{A model fit as produced by \link{\code{fit_tmb}}}

\item{type}{The type of biomass to plot, options are 'ssb',
'total', or 'summary' which are spawning, total, and age 2+,
respectively. Default is SSB.}

\item{plot}{Whether to create a plot and invisibly return the
ggplot object,  or return a data.frame of the data.}
}
\value{
A ggplot or data frame depending on \code{plot}.
}
\description{
Plot cohort contribution to SSB as a stacked ribbon plot
}
\details{
The plus group breaks the idea of a cohort a little
bit. Here it assumes that the cohort dies when hitting the
plus group.
}
