% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fns.R
\name{plot_data_overview}
\alias{plot_data_overview}
\title{Plot overview of data sources}
\usage{
plot_data_overview(datlist)
}
\arguments{
\item{datlist}{Datlist as returned by \code{read_dat}.}
}
\value{
Makes plot and invisibly returns the processed tidy
data frame
}
\description{
Plot overview of data sources
}
\details{
This works well with png(..., width=7, height=5,
units='in')
}
