% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fns.R
\name{plot_ess}
\alias{plot_ess}
\title{Plot effective sample sizes to compare Francis tuning to Dirichlet-multinomial (DM)}
\usage{
plot_ess(fitDM, fitfrancis = NULL, addISS = FALSE, plot = TRUE)
}
\arguments{
\item{fitDM}{A model fit with the DM turned on}

\item{fitfrancis}{A model fit with Francis tuning}

\item{addISS}{Whether to add the input sample size (ISS) taken from the fitDM model.}

\item{plot}{Whether to print.}
}
\value{
A ggplot object if plot is TRUE, otherwise the data
}
\description{
Plot effective sample sizes to compare Francis tuning to Dirichlet-multinomial (DM)
}
