% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fns.R
\name{plot_index_fits}
\alias{plot_index_fits}
\title{Plot fits to biomass indices for a single fit}
\usage{
plot_index_fits(fit)
}
\arguments{
\item{fit}{A fitted model object}
}
\description{
Plot fits to biomass indices for a single fit
}
