% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fns.R
\name{plot_obs_exp}
\alias{plot_obs_exp}
\title{Plot observed vs expected compositions as lines}
\usage{
plot_obs_exp(mat, years, ncol = 5, title, minyr = NULL, minage = NULL)
}
\arguments{
\item{mat}{Matrix of observed and expected compositions as
returned by \code{read_pk_rep}}

\item{years}{Vector of years with data}

\item{ncol}{Number of columns for faceting}

\item{title}{Plot title}

\item{minyr}{Optional minimum year to plot}

\item{minage}{Optional minimum age to plot}
}
\value{
Prints a ggplot
}
\description{
Plot observed vs expected compositions as lines
}
