% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osa_comp_plots.R
\name{plot_osa_comps}
\alias{plot_osa_comps}
\title{Explore OSA residuals for multinomial composition data and
compare to Pearson}
\usage{
plot_osa_comps(
  obs,
  exp,
  ages,
  years,
  Neff,
  model,
  plot.type = c("default", "ggplot")
)
}
\arguments{
\item{obs, exp, pearson}{the observed, expected and Pearson
residual matrices with rows as years and columns as ages (or
lengths)}

\item{ages, years}{vectors giving the ages and years}

\item{model}{Character for model version}
}
\value{
returns nothing but creates a PDF file in the working
directory
}
\description{
Explore OSA residuals for multinomial composition data and
compare to Pearson
}
