% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fns.R
\name{plot_osa_fits}
\alias{plot_osa_fits}
\title{Plot OSA fits via afscOSA for fishery and all surveys}
\usage{
plot_osa_fits(fit)
}
\arguments{
\item{fit}{Fitted object}
}
\description{
Plot OSA fits via afscOSA for fishery and all surveys
}
\details{
This is just a wrapper around run_osa
}
