% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fns.R
\name{plot_pk_ssb}
\alias{plot_pk_ssb}
\title{Plot spawning biomass}
\usage{
plot_pk_ssb(
  x,
  add_uncertainty = TRUE,
  plotlog = TRUE,
  uselog = FALSE,
  plot = TRUE
)
}
\arguments{
\item{x}{A list of data frames as read in from
\link{\code{read_pk_cor}}}

\item{add_uncertainty}{Whether to add 95\% confidence
intervals. These are calculated in log space.}

\item{plotlog}{Whether to plot in log space or not}

\item{uselog}{Whether to calculate in log space or not}

\item{plot}{Whether to plot and return the ggplot object
(TRUE) or return the data (FALSE).}
}
\value{
Either a ggplot object or the data used for it,
depending on \code{plot} argument.
}
\description{
Plot spawning biomass
}
