% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fns.R
\name{plot_resids}
\alias{plot_resids}
\title{Plot Pearson residual matrix}
\usage{
plot_resids(mat, years, minyr = NULL)
}
\arguments{
\item{mat}{Matrix of Pearson residuals (years by ages) as
returned by \code{read_pk_rep}}

\item{years}{Vector of years of observations}
}
\value{
A long data frame ready for ggplot
}
\description{
Plot Pearson residual matrix
}
