% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retro_fns.R
\name{plot_retro_recruits}
\alias{plot_retro_recruits}
\title{Plot estimates of recruits from retro lists}
\usage{
plot_retro_recruits(fits, thisyear, plot = TRUE)
}
\arguments{
\item{fits}{List of retro fits including peel=0}

\item{thisyear}{The current year}

\item{plot}{Whether to plot or return}
}
\description{
Plot estimates of recruits from retro lists
}
