% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fns.R
\name{plot_retros}
\alias{plot_retros}
\title{Plot retrospective peels against the base model with CI}
\usage{
plot_retros(retros, type = c("ssb", "F", "recruit"), title = NULL)
}
\arguments{
\item{retros}{A list of pkfits}

\item{type}{The metric to use: 'ssb', 'F', 'recruit'}
}
\description{
Plot retrospective peels against the base model with CI
}
