% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb_fns.R
\name{prepare_par}
\alias{prepare_par}
\title{Internal fucntion to prepare parameter list from dat
list. Uses initial values from the 2022 model for scalars, and
vectors are all 0.}
\usage{
prepare_par(dat)
}
\description{
Internal fucntion to prepare parameter list from dat
list. Uses initial values from the 2022 model for scalars, and
vectors are all 0.
}
