% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb_fns.R
\name{prepare_pk_input}
\alias{prepare_pk_input}
\title{Prepare inputs for the TMB GOA pollock model.}
\usage{
prepare_pk_input(dat, version = "none", random = NULL)
}
\arguments{
\item{dat}{Data list as read in with
\link{\code{read_pk_dat}}.}

\item{version}{A character string for the model name, used for
plotting downstream}

\item{random}{A character vector declaring random effect
vectors to be integrated. Defaults to NULL (fully penalized
ML)}
}
\value{
A list with the version, dat, pars, map and random
which are used to build a TMB 'obj' in
\link{\code{fit_pk}}.
}
\description{
Prepare inputs for the TMB GOA pollock model.
}
