% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.pkfit}
\alias{print.pkfit}
\title{Print summary of pkfit object}
\usage{
\method{print}{pkfit}(fit, ...)
}
\arguments{
\item{fit}{Fitted object from \code{\link{fit_pk}}}

\item{...}{Ignored for now}
}
\value{
Summary printed to console
}
\description{
Print summary of pkfit object
}
