% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admb_fns.R
\name{read_dat}
\alias{read_dat}
\title{Read text file input for ADMB model}
\usage{
read_dat(filename, path = NULL)
}
\arguments{
\item{filename}{The filename to be read in}

\item{path}{Path to the file if not in working directory}
}
\description{
Read text file input for ADMB model
}
