% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pk_rep.R
\name{read_pk_cor}
\alias{read_pk_cor}
\title{Read correlation file}
\usage{
read_pk_cor(
  model = "goa_pk",
  path = getwd(),
  version = "none",
  endyr,
  styr = 1970
)
}
\arguments{
\item{model}{model name}

\item{path}{path to folder}

\item{version, endyr, styr}{see rep function}
}
\description{
Read correlation file
}
