% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_fns.R
\name{read_pk_rep}
\alias{read_pk_rep}
\title{Read the GOA pollock model output report}
\usage{
read_pk_rep(
  model = "goa_pk",
  path = getwd(),
  version = "none",
  endyr,
  styr = 1970
)
}
\arguments{
\item{model}{model name}

\item{path}{path to folder}

\item{version}{A version name which is added, e.g., 'change_selex'}

\item{endyr, styr}{The start and end years in the model}
}
\value{
A list of outputs
}
\description{
Read the GOA pollock model output report
}
