% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admb_fns.R
\name{read_pk_std}
\alias{read_pk_std}
\title{Read standard deviation file}
\usage{
read_pk_std(
  model = "goa_pk",
  path = getwd(),
  version = "none",
  endyr,
  styr = 1970
)
}
\arguments{
\item{model}{model name}

\item{path}{path to folder}

\item{version, endyr, styr}{see rep function}
}
\description{
Read standard deviation file
}
