% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_francis_weighting}
\alias{run_francis_weighting}
\title{Run Francis weighting on a fitted model}
\usage{
run_francis_weighting(fit, iter = 10, print.check = TRUE)
}
\arguments{
\item{fit}{A fitted object returned by \code{fit_pk}}

\item{iter}{The number of iterations to do, defaults to 10}

\item{print.check}{Whether to print a plot showing the weights
by iteration for each survey}
}
\value{
A fitted object using from the last iteration
}
\description{
Run Francis weighting on a fitted model
}
\details{
Implements Francis weighting for age composition data
except for survey 3 which needs to be fixed still.
}
