% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb_fns.R
\name{run_jitter}
\alias{run_jitter}
\title{Run a jitter analysis from a fitted model}
\usage{
run_jitter(fit, njitter, scalar = 0.1, parallel = TRUE)
}
\arguments{
\item{njitter}{}
}
\description{
Run a jitter analysis from a fitted model
}
