% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admb_fns.R
\name{setup_exe}
\alias{setup_exe}
\title{Copy and build ADMB source from package}
\usage{
setup_exe(path = getwd(), name = "goa_pk", compile = TRUE)
}
\arguments{
\item{path}{Path to folder}

\item{name}{Executable name (default goa_pk)}

\item{compile}{Whether to compile (default is TRUE)}
}
\description{
Copy and build ADMB source from package
}
