% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fns.R
\name{sim_dat}
\alias{sim_dat}
\title{Simulate a data set from a given rep file and input data}
\usage{
sim_dat(datlist, replist, fileout = NULL, path = NULL)
}
\arguments{
\item{datlist}{A data list as returned by
\link{read_dat}. Observed data overwritten with simulated
data.}

\item{replist}{A replist as returned by
\link{read_pk_rep}. Uses expected values as truth.}

\item{fileout}{The new dat file name to be written}

\item{path}{Optional path for fileout}
}
\value{
Nothing, a data file is written.
}
\description{
Simulate a data set from a given rep file and input data
}
\details{
The expected values from the report file are used as
the "truth." This routine is similar to the approach used in
SS.Data simulated either as lognormal (indices) or
multinomial (compositions) given the CV and sample sizes
given in the datlist.  This implicitly assumes they have
been weighted as desired. Sample sizes of 0 result in no
observations, and fractions will be rounded internally by
rmultinom. Currently the total fishery catch is not
resampled, but the age compositions are.
}
