% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{surveyf}
\alias{surveyf}
\title{Create factor from numeric survey vector. Useful for
informative plots and tables}
\usage{
surveyf(x)
}
\arguments{
\item{x}{Vector of survey numbers, 1,2, etc. which are
transformed into a named factor}
}
\value{
Vector of named surveys
}
\description{
Create factor from numeric survey vector. Useful for
informative plots and tables
}
