% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb_fns.R
\name{update_dat_maxage}
\alias{update_dat_maxage}
\title{Update a data list with a new maximum age, summing age compositions, but truncating WAA, ageing error, length transition matrices, and maturity.}
\usage{
update_dat_maxage(dat, maxage)
}
\arguments{
\item{dat}{A data list as returned by \code{read_dat}}

\item{maxage}{A new maximum age that is less than the one in the dat file.}
}
\value{
An updated data input list that is reformatted to the new maximum age.
}
\description{
Update a data list with a new maximum age, summing age compositions, but truncating WAA, ageing error, length transition matrices, and maturity.
}
