% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fns.R
\name{write_dat}
\alias{write_dat}
\title{Write an R data list to file for reading into ADMB}
\usage{
write_dat(datlist, fileout, path = NULL)
}
\arguments{
\item{datlist}{An object as read in by \link{read_dat}}

\item{fileout}{The new filename to be written}

\item{path}{Optional path if not in the working directory}
}
\value{
Nothing, new file written to disk
}
\description{
Write an R data list to file for reading into ADMB
}
