% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_fns.R
\name{write_proj_inputs}
\alias{write_proj_inputs}
\title{Write input files for the Projection module}
\usage{
write_proj_inputs(replist, datlist, path = getwd())
}
\arguments{
\item{replist}{A list as read in by \link{read_rep}}

\item{datlist}{A list as read in by \link{read_dat}}

\item{path}{Directory to write in}
}
\value{
Nothing but writes three files
}
\description{
Write input files for the Projection module
}
