% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_fns.R
\name{write_spm_inputs}
\alias{write_spm_inputs}
\title{Write input files for the 'spm' projection module. Uses subsequent year's NAA
(endN) to better match the assessment model dynamics of the terminal year.}
\usage{
write_spm_inputs(fit, path = NULL)
}
\arguments{
\item{fit}{A fitted object from \code{fit_pk}.}

\item{path}{Directory to write output files to}
}
\value{
Nothing but writes files spm.dat, goa_wp.txt, and tacpar.dat (dummy file)
to 'path' based on values in the replist and datlist objections
}
\description{
Write input files for the 'spm' projection module. Uses subsequent year's NAA
(endN) to better match the assessment model dynamics of the terminal year.
}
