% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_fns.R
\name{calculate_rho}
\alias{calculate_rho}
\title{Calculate Mohn's rho}
\usage{
calculate_rho(reps, max_peels = NULL)
}
\arguments{
\item{reps}{A list of replists for each peel as returned by
\code{read_pk_rep}}

\item{max_peels}{Optional numeric to control the max number of
peels to use in the calculation. E.g. can calculate rho with
4,5,6, etc. peels and compare}
}
\value{
The rho value rounded to 3 decimal places, and prints
to console
}
\description{
Calculate Mohn's rho
}
