% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_fns.R
\name{get_exec_table}
\alias{get_exec_table}
\title{Get table from projection output}
\usage{
get_exec_table(replist, bigfile)
}
\arguments{
\item{replist}{A list as read in by \link{read_pk_rep}}

\item{bigfile}{A data frame of full projection scenario
outputs}
}
\description{
Get table from projection output
}
