% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_fns.R
\name{get_exec_table}
\alias{get_exec_table}
\title{Get table from projection output}
\usage{
get_exec_table(fit, summary, detail, maxABCratio = 1)
}
\arguments{
\item{fit}{A model fit with \code{fit_pk}}

\item{summary}{The spm_summary.csv object}

\item{detail}{The spm_detail.csv object}
}
\description{
Get table from projection output
}
