% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb_fns.R
\name{get_rep}
\alias{get_rep}
\title{Extract report list from fitted models}
\usage{
get_rep(fits)
}
\arguments{
\item{fits}{A single model or list of models as returned by
\code{\link{fit_pk}}}
}
\value{
Reports
}
\description{
Extract report list from fitted models
}
