% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osa_comp_plots.R, R/output_fns.R
\name{plot_osa_comps}
\alias{plot_osa_comps}
\title{Explore OSA residuals for multinomial composition data and
compare to Pearson}
\usage{
plot_osa_comps(
  res,
  years,
  Neff,
  ind = 10,
  survey,
  drop = NULL,
  plot = TRUE,
  minyear = NULL,
  model = c("multinomial", "Dirichlet-multinomial"),
  theta
)

plot_osa_comps(
  res,
  years,
  Neff,
  ind = 10,
  survey,
  drop = NULL,
  plot = TRUE,
  minyear = NULL,
  model = c("multinomial", "Dirichlet-multinomial"),
  theta
)
}
\arguments{
\item{res}{A matrix of obs and exp results cbinded together,
as returned by the model}

\item{years}{Vector of years for plotting}

\item{Neff}{Effective sample size vector. Non-integer values
are rounded currently.}

\item{ind}{Which age to drop, modified as need when 'drop' is
used.}

\item{survey}{Character name for plotting}

\item{drop}{Optional vector of ages to drop. Alters meaning of
'ind'}

\item{plot}{Whether to plot or return the data instead}

\item{minyear}{Filter out years before this}

\item{model}{Currently supports 'multinomial' and
'Dirichlet-multinomial', when using hte latter the parameter
theta must be passed to recalculate the "alpha" term
internally.}

\item{theta}{A scalar parameter estimated as exp(log_DM_pars)
corresponding to this survey}

\item{obs, exp, pearson}{the observed, expected and Pearson
residual matrices with rows as years and columns as ages (or
lengths)}

\item{ages, years}{vectors giving the ages and years}
}
\value{
returns nothing but creates a PDF file in the working
directory

A ggplot object if plot is TRUE or the melted data if
FALSE
}
\description{
Explore OSA residuals for multinomial composition data and
compare to Pearson

Plot OSA composition residuals as bubble plots
}
