% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fns.R
\name{plot_pk_selex}
\alias{plot_pk_selex}
\title{Plot survey selectivities}
\usage{
plot_pk_selex(x, add_uncertainty = TRUE, plot_logit = FALSE, plot = TRUE)
}
\arguments{
\item{x}{A list of data frames as read in from
\link{\code{read_pk_std}}}

\item{add_uncertainty}{Whether to add 95\% confidence
intervals. These are calculated in logit space.}

\item{plot_logit}{Whether to plot in logit space or not}

\item{plot}{Whether to plot and return the ggplot object
(TRUE) or return the data (FALSE).}
}
\value{
Either a ggplot object or the data used for it,
depending on \code{plot} argument.
}
\description{
Plot survey selectivities
}
