% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fns.R
\name{plot_pk_ssb}
\alias{plot_pk_ssb}
\title{Plot spawning biomass}
\usage{
plot_pk_ssb(
  fits,
  add_uncertainty = TRUE,
  plotlog = FALSE,
  uselog = FALSE,
  addproj = FALSE,
  plot = TRUE,
  alpha1 = 0.5,
  alpha2 = 0.8
)
}
\arguments{
\item{fits}{A model fit or list of model fits as produced by
\link{\code{fit_tmb}}}

\item{add_uncertainty}{Whether to add 95\% confidence
intervals. These are calculated in log space depending on
\code{uselog}.}

\item{plotlog}{Whether to plot SSB in log space or not}

\item{uselog}{Whether to calculate in log space or not}

\item{plot}{Whether to plot and return the ggplot object
(TRUE) or return the data used in the plot(FALSE).}

\item{alpha1}{Transparency for ribbons}

\item{alpha2}{Transparency for lines}
}
\value{
Either a ggplot object or the data used for it,
depending on \code{plot} argument.
}
\description{
Plot spawning biomass
}
