% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb_fns.R
\name{prepare_pk_input}
\alias{prepare_pk_input}
\title{Prepare inputs for the TMB GOA pollock model.}
\usage{
prepare_pk_input(
  path,
  datfile,
  version = "none",
  random = NULL,
  complike = c("multinomial", "D-M"),
  modfile = "goa_pk",
  maxage = NULL
)
}
\arguments{
\item{path}{Directory containing the model and dat files.
Passed to \code{fit_pk}.}

\item{datfile}{Name of dat file to be read in.}

\item{version}{A character string for the model name, used for
plotting downstream}

\item{random}{A character vector declaring random effect
vectors to be integrated. Defaults to NULL (fully penalized
ML)}

\item{complike}{A flag to use either 'multinomial' or
Dirichlet-multinomial 'D-M'.}

\item{modfile}{Model name assumed to be 'goa_pk' unless
specified}

\item{maxage}{Specify a lower maximum age than in the input
dat file if desired.}
}
\value{
A list with the version, dat, pars, map and random
which are used to build a TMB 'obj' in \code{\link{fit_pk}}.
}
\description{
Prepare inputs for the TMB GOA pollock model.
}
