% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb_fns.R
\name{read_dat}
\alias{read_dat}
\title{Read text file input for model. Excel sheet is saved as .txt and read into R with this function.}
\usage{
read_dat(filename, path = NULL, writedat = FALSE)
}
\arguments{
\item{filename}{The filename to be read in}

\item{path}{Path to the file if not in working directory}
}
\value{
A named list of all elements with corresponding names
to the ADMB model
}
\description{
Read text file input for model. Excel sheet is saved as .txt and read into R with this function.
}
