% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pk_rep.R
\name{read_pk_rep}
\alias{read_pk_rep}
\title{Read the GOA pollock model output report}
\usage{
read_pk_rep(file, endyr = 2020, styr = 1970, version = "none")
}
\arguments{
\item{file}{Report file name}

\item{endyr, styr}{The start and end years in the model}

\item{version}{A version name which is added, e.g., 'change_selex'}
}
\value{
A list of outputs
}
\description{
Read the GOA pollock model output report
}
