% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb_fns.R
\name{run_jitter}
\alias{run_jitter}
\title{Run a jitter analysis on a fitted model}
\usage{
run_jitter(fit, njitter = 20, scalar = 0.1, parallel = TRUE, type = "par")
}
\arguments{
\item{fit}{A fitted model object from \code{fit_pk}.}

\item{njitter}{The number of jitter runs to do.}

\item{scalar}{How much to scale the values, where a new init value is generated by multiplying by U(1-scalar,1+scalar)}

\item{parallel}{Whether to run in parallel (default and recommended).}

\item{type}{Either "mle" which jitters around the MLE value or "par" where the initial values in obj$par are used.}
}
\value{
A data.frame with columns for the jitter number, terminal SSB, parameter inits and estimations, and maximum gradient. It also prints to console
}
\description{
Run a jitter analysis on a fitted model
}
