% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fns.R
\docType{data}
\name{sim_dat}
\alias{sim_dat}
\title{Simulate a data set from a given rep file and input data}
\format{
An object of class \code{data.frame} with 225 rows and 4 columns.
}
\usage{
sim_dat
}
\arguments{
\item{datlist}{A data list as returned by
\link{read_dat}. Observed data overwritten with simulated
data.}

\item{replist}{A replist as returned by
\link{read_pk_rep}. Uses expected values as truth.}

\item{fileout}{The new dat file name to be written. If NULL no
file is written.}

\item{path}{Optional path for fileout}

\item{type}{Either "model" which uses the model expectation as
the means, or "data" which uses the observed data means.}
}
\value{
An invisible list is returned, and optionally a file
written. The age compositions are returned as proportions
because the model assumes this.
}
\description{
Simulate a data set from a given rep file and input data
}
\details{
The expected values from either the dat or report
file are used as the "truth" depending on \code{type}. The
"model" routine is similar to the approach used in SS.Data
simulated either as lognormal (indices) or multinomial
(compositions) given the CV and sample sizes given in the
datlist.  This implicitly assumes they have been weighted as
desired. The "data" approach is akin to a non-parametric
bootstrap. Sample sizes of 0 result in no observations, and
fractions will be rounded internally by rmultinom. Currently
the total fishery catch is not resampled, but the age
compositions are.
}
\keyword{datasets}
