% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_fns.R
\name{write_spm_inputs}
\alias{write_spm_inputs}
\title{Write input files for the 'spm' projection module. Differs
slightly from the old projection module.}
\usage{
write_spm_inputs(replist, datlist, path = getwd())
}
\arguments{
\item{replist}{A list as read in by \link{read_rep}}

\item{datlist}{A list as read in by \link{read_dat}}

\item{path}{Directory to write in}
}
\value{
Nothing but writes files spm.dat, goa_wp.txt, and
tacpar.dat (dummy file) to 'path' based on values in the
replist and datlist objections
}
\description{
Write input files for the 'spm' projection module. Differs
slightly from the old projection module.
}
